removed need for extra channel being passed to service
